Use abbot-cli to manage the sql skill
