+++ sql/.meta/.language
@@ -1,0 +1,1 @@
+﻿Python
